pyproject.toml
[project]
name  = "db_tools"
version = "0.1.0"
description = "Handler for multiple database connections"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
